Add /slim/lead/:title route
Completely move the slim transform into common file so that it can be
shared with the slim-lead transform.

Add the slim-lead transform, which simply proxies to the slim transform,
and returns the first section in the result.

This addresses #18 but doesn't close it as references for the lead
section aren't returned.