Make imagify maintain the original node
Instead of creating a new one and putting the old node's html in an attribute.

Now it just removes the src and srcset of the images and moves them to the
data-src and data-srcset attributes for the client to replace back. It also
sets a title so that it shows when the image is not loaded, and some inline
styles to preserve layout when the image is not loaded.